+++ sphinx/source/changelog.rst
@@ -8,6 +8,8 @@ Changelog
 * Added the :ref:`defw` directive (for inserting word values into the memory
   snapshot)
 * Added the :ref:`remote` directive (for creating a remote entry)
+* Added the ``--poke`` option to :ref:`bin2sna.py` (for performing POKE
+  operations on the snapshot)
 * Added support to the :ref:`Resources` section for specifying files using
   wildcard characters (``*``, ``?`` and ``[]``)
 * Added support for parsing block-level comments that are not left-padded by a